Adds draft of template for starting context model testing.
