+++ hrc_speech_prediction/tests/test_combined_model.py
@@ -3,23 +3,10 @@ from unittest import TestCase
 
 from sklearn.linear_model import SGDClassifier
 
-from hrc_speech_prediction import context_model
-from hrc_speech_prediction.data import (ALL_ACTIONS, TRAIN_PARTICIPANTS,
-                                        TrainData)
-from hrc_speech_prediction.defaults import MODEL_PATH
+from hrc_speech_prediction.data import (ALL_ACTIONS, TrainData)
 from hrc_speech_prediction.features import get_bow_features
 from hrc_speech_prediction.models import CombinedModel
 from hrc_speech_prediction.speech_model import SpeechModel
-
-
-class TestContextTreeModel(TestCase):
-    def setUp(self):
-        self.model = context_model.ContextTreeModel(ALL_ACTIONS)
-
-        self.model.fit([[], [], ["a"], ["z", "b", "c"]], ["a", "a", "b", "d"])
-
-    def test_n_children(self):
-        self.assertEqual(self.model.root.n_children, 2)
 
 
 class TestCombinedModel(TestCase):
